SBERDOMA-1122 moved `addressMeta` validation utils to "property" domain
