--- apps/condo/domains/common/utils/addressApi/AddressMetaSchema.js
@@ -1,37 +1,0 @@
-const { AddressMetaDataFields } = require('../../../property/schema/fields/AddressMetaField')
-
-const AddressMetaJSONSchema = {
-    'type': 'object',
-    'properties': {
-        'value': {
-            'type': 'string',
-        },
-        'unrestricted_value': {
-            'type': 'string',
-        },
-        'dv': {
-            'type': 'integer',
-        },
-        'data': {
-            'type': 'object',
-            'properties': Object.assign({},
-                ...Object.keys(AddressMetaDataFields).map((x) => ({ [x]: { 'type': ['string', 'null'] } })),
-                { history_values: { type: ['array', 'null'], items: { type: 'string' } } }
-            ),
-            'additionalProperties': true,
-            'required': Object.keys(AddressMetaDataFields),
-        },
-    },
-    'additionalProperties': true,
-    'required': [
-        'value',
-        'unrestricted_value',
-        'data',
-        'dv',
-    ],
-}
-
-module.exports = {
-    AddressMetaJSONSchema,
-    AddressMetaDataFields,
-}
