+++ apps/condo/domains/property/schema/fields/AddressMetaField.js
@@ -1,3 +1,5 @@
+const Ajv = require('ajv')
+
 const AddressMetaDataFields = {
     postal_code: 'String!',
     country: 'String!',
@@ -115,10 +117,46 @@ const ADDRESS_META_FIELD_GRAPHQL_TYPES = `
     }
 `
 
+const AddressMetaJSONSchema = {
+    'type': 'object',
+    'properties': {
+        'value': {
+            'type': 'string',
+        },
+        'unrestricted_value': {
+            'type': 'string',
+        },
+        'dv': {
+            'type': 'integer',
+        },
+        'data': {
+            'type': 'object',
+            'properties': Object.assign({},
+                ...Object.keys(AddressMetaDataFields).map((x) => ({ [x]: { 'type': ['string', 'null'] } })),
+                { history_values: { type: ['array', 'null'], items: { type: 'string' } } }
+            ),
+            'additionalProperties': true,
+            'required': Object.keys(AddressMetaDataFields),
+        },
+    },
+    'additionalProperties': true,
+    'required': [
+        'value',
+        'unrestricted_value',
+        'data',
+        'dv',
+    ],
+}
+
+
+const ajv = new Ajv()
+const addressMetaJsonValidator = ajv.compile(AddressMetaJSONSchema)
+
 const ADDRESS_META_SUBFIELDS_QUERY_LIST = `dv value unrestricted_value data { ${Object.keys(AddressMetaDataFields).join(' ')} }`
 
 module.exports = {
     ADDRESS_META_FIELD_GRAPHQL_TYPES,
     ADDRESS_META_SUBFIELDS_QUERY_LIST,
-    AddressMetaDataFields
+    AddressMetaDataFields,
+    addressMetaJsonValidator,
 }
